Splitting up regression models.
